Fix build dependency on Mac
Upstream introduced a new build step, which has a requirement to the `generate_symlinks` target.
This target is declared conditionally only when the build is not on the target host.

It causes the following error:

```
ERROR Unresolved dependencies.
//flutter/shell/platform/embedder:flutter_embedder_framework_archive(//build/toolchain/mac:clang_x64)
  needs //flutter/shell/platform/embedder:generate_symlinks(//build/toolchain/mac:clang_x64)
```

This commit introduces a patch on top of the file, which moves the new target
to be declared conditionally to match the same rules. Ideally, we can move it
upstream to avoid needing this patch - but I'm doing this on this repo to
restabilish the build stage so we can continue using it meanwhile.